refactor(gateway): extract sandbox init into prepare_core/sandbox.rs
Move sandbox router construction and background task spawning (image
pre-build, host provisioning, container GC) from prepare_core.rs into
a dedicated submodule. Reduces prepare_core.rs from 1503 to 1286 lines,
fixing the CI file-size check.

New module `prepare_core/sandbox.rs` exposes two functions:
- `build_sandbox_router()` — creates router with remote backends registered
- `spawn_sandbox_background_tasks()` — image build, host provision, GC
